Former-commit-id: 8f65c2152a8d8bc2ef3d68afa2bc1495ea3e463c

+++ dao/base.py
@@ -1,4 +1,6 @@
+import csv
 import datetime
+import fileinput
 
 from sqlalchemy import create_engine, Column, Integer, String, DateTime
 from sqlalchemy.ext.declarative import declarative_base
@@ -28,15 +30,19 @@ class Connection(metaclass=Singleton):
         self.create_db()
 
     def drop_db(self):
-        self.Base.metadata.drop_all(self.engine)
-        print("DB dropped")
+        resp = input('Please type yes if you would like to drop the whole database')
+        if str(resp).lower() == 'yes':
+            self.Base.metadata.drop_all(self.engine)
+            print("DB dropped")
+        else:
+            raise Exception('Db drop failed, canceling further steps')
 
     def create_db(self):
         self.Base.metadata.create_all(self.engine)
         print("DB created")
 
     def open_session(self):
-        print("Opening new session")
+        # print("Opening new session")
         return self.Session()
